facial_drooping/.other/2/requirements.txt
Flask==2.1.2
tensorflow==2.10.0
Pillow==9.1.1
numpy==1.22.4
